Yellow Fever Data from Brazil; 2016-12 to 2017-05

This data set contains flows to and from five states in Brazil formatted in a list with the following items:

data("Brazil_epiflows")
  data("YF_coordinates")
  data("YF_locations")
  data("YF_flows")
  data("YF_Brazil")

Format

An object of class list of length 4.

Details

$states: a data frame containing metadata for five Brazilian States: Espirito Santo, Minas Gerais, Rio de Janeiro, Sao Paulo, and Southeast Brazil

$location_code : names of the states

$location_population : population size for each state

$num_cases_time_window : number of cases recorded between 2016-12 and 2017-05

$first_date_cases : date of first disease case in the given location in ISO 8601 format

$last_date_cases : date of last disease case in the given location in ISO 8601 format

$T_D A matrix containing the number of travellers from the infectious location visiting other locations

$T_O A matrix containing the number of travellers visiting the infectious location

$length_of_stay A named vector containing the average length of stay in days of travellers from other locations visiting the infectious locations.

References

Dorigatti I, Hamlet A, Aguas R, Cattarino L, Cori A, Donnelly CA, Garske T, Imai N, Ferguson NM. International risk of yellow fever spread from the ongoing outbreak in Brazil, December 2016 to May 2017. Euro Surveill. 2017;22(28):pii=30572. DOI: 10.2807/1560-7917.ES.2017.22.28.30572

See also

Examples

# This is an example of an epiflows object
data("Brazil_epiflows")
Brazil_epiflows

#> 
#> /// Epidemiological Flows //
#> 
#>   // class: epiflows, epicontacts
#>   // 15 locations; 100 flows; directed
#>   // optional variables: pop_size, duration_stay, num_cases, first_date, last_date 
#> 
#>   // locations
#> 
#> # A tibble: 15 x 6
#>    id    location_popula… num_cases_time_… first_date_cases last_date_cases
#>  * <chr>            <dbl>            <dbl> <fct>            <fct>          
#>  1 Espi…          3973697             2600 2017-01-04       2017-04-30     
#>  2 Mina…         20997560             4870 2016-12-19       2017-04-20     
#>  3 Rio …         16635996              170 2017-02-19       2017-05-10     
#>  4 Sao …         44749699              200 2016-12-17       2017-04-20     
#>  5 Sout…         86356952             7840 2016-12-17       2017-05-10     
#>  6 Arge…               NA               NA <NA>             <NA>           
#>  7 Chile               NA               NA <NA>             <NA>           
#>  8 Germ…               NA               NA <NA>             <NA>           
#>  9 Italy               NA               NA <NA>             <NA>           
#> 10 Para…               NA               NA <NA>             <NA>           
#> 11 Port…               NA               NA <NA>             <NA>           
#> 12 Spain               NA               NA <NA>             <NA>           
#> 13 Unit…               NA               NA <NA>             <NA>           
#> 14 Unit…               NA               NA <NA>             <NA>           
#> 15 Urug…               NA               NA <NA>             <NA>           
#> # ... with 1 more variable: length_of_stay <dbl>
#> 
#>   // flows
#> 
#> # A tibble: 100 x 3
#>    from             to         n
#>    <chr>            <chr>  <dbl>
#>  1 Espirito Santo   Italy  2828.
#>  2 Minas Gerais     Italy 15714.
#>  3 Rio de Janeiro   Italy  8164.
#>  4 Sao Paulo        Italy 34039.
#>  5 Southeast Brazil Italy 76282.
#>  6 Espirito Santo   Spain  3270.
#>  7 Minas Gerais     Spain 18176.
#>  8 Rio de Janeiro   Spain  9443.
#>  9 Sao Paulo        Spain 39371.
#> 10 Southeast Brazil Spain 88231.
#> # ... with 90 more rows
#> 

# The above data was constructed from a data frame containing flows and 
# one containing location metadata
data("YF_flows")
data("YF_locations")
ef <- make_epiflows(flows         = YF_flows,
                    locations     = YF_locations,
                    pop_size      = "location_population",
                    duration_stay = "length_of_stay",
                    num_cases     = "num_cases_time_window",
                    first_date    = "first_date_cases",
                    last_date     = "last_date_cases"
                   )

# Both of the above data frames were constructed like so:

data("YF_Brazil")

# Create the flows data frame
from  <- as.data.frame.table(YF_Brazil$T_D, stringsAsFactors = FALSE)
to    <- as.data.frame.table(t(YF_Brazil$T_O), stringsAsFactors = FALSE)
flows <- rbind(from, to)
colnames(flows) <- c("from", "to", "n")

## Create the locations data frame
los <- data.frame(location_code    = names(YF_Brazil$length_of_stay),
                  length_of_stay   = YF_Brazil$length_of_stay,
                  stringsAsFactors = FALSE
                 )
locations <- merge(x   = YF_Brazil$states,
                   y   = los,
                   by  = "location_code",
                   all = TRUE)

## Use both to create the epiflows object.
ef <- make_epiflows(flows,
                    locations,
                    pop_size = "location_population",
                    duration_stay = "length_of_stay",
                    num_cases = "num_cases_time_window",
                    first_date = "first_date_cases",
                    last_date = "last_date_cases"
)